Move DeleteRequest class to common/deletions module
