+++ gear/form_deletion/src/python/form_deletion_app/run.py
@@ -4,6 +4,7 @@ import logging
 from typing import Optional
 
 from configs.ingest_configs import FormProjectConfigs, load_form_ingest_configurations
+from deletions.models import DeleteRequest
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor, ProjectError
 from fw_gear import GearContext
 from gear_execution.gear_execution import (
@@ -19,7 +20,6 @@ from identifiers.model import IdentifiersMode
 from inputs.parameter_store import ParameterStore
 from lambdas.lambda_function import LambdaClient, create_lambda_client
 from pydantic import ValidationError
-from submissions.models import DeleteRequest
 
 from form_deletion_app.main import run
